organized student form specific components into folder
